requirements.txt
# ARCHIVED: We are transitioning to pyproject.toml and uv-based project management
# However, we provide this as a backup for now

# Frameworks
# we use latest PyTorch stable release
torch==2.9.*
triton==3.5.*

# we shall upgrade torch for blackwell when it is stable
transformers>=4.57.3
datasets>=4.4.2
modal>=1.3.0

# DSLs
nvidia-cutlass-dsl
tilelang
<<<<<<< Updated upstream
=======
triton
cuda-tile  # cuTile - requires CUDA Toolkit 13.1+
>>>>>>> Stashed changes

# helper
tqdm>=4.67.1
packaging
pydra-config
ninja>=1.13.0
cupy-cuda12x==13.6.0
tomli>=2.3.0
tabulate>=0.9.0
nsight-python

# Numerics
einops>=0.8.1
python-dotenv>=1.2.1
numpy==2.4.0

# use litellm for cloud providers and openai for local
openai>=2.14.0
litellm[proxy]>=1.80.10